tg_iki_bot/requirements.txt
python-telegram-bot==20.7 